requirements.txt
asgiref==3.8.1
Django==4.2.13
djangorestframework==3.15.1
numpy==1.24.3
pandas==2.0.1
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.0
tzdata==2023.3
asgiref==3.8.1
asttokens==2.4.1
asyncio==3.4.3
certifi==2024.6.2
chardet==4.0.0
colorama==0.4.6
decorator==5.1.1
Django==4.2.13
djangorestframework==3.15.1
exceptiongroup==1.2.1
executing==2.0.1
fyers-apiv2==2.0.9
idna==2.10
ipython==8.18.1
jedi==0.19.1
matplotlib-inline==0.1.7
numpy==1.24.3
pandas==2.0.1
parso==0.8.4
prompt_toolkit==3.0.45
psycopg2-binary==2.9.9
pure-eval==0.2.2
Pygments==2.18.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.25.1
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
tornado==6.1
traitlets==5.14.3
typing_extensions==4.12.0
tzdata==2023.3
urllib3==1.26.18
wcwidth==0.2.13
websocket-client==1.2.1
websockets==8.1
